Remove all traces of dagger, hilt and kapt
