+++ app/src/main/kotlin/net/marcoromano/mooviez/app/Application.kt
@@ -1,7 +1,5 @@
 package net.marcoromano.mooviez.app
 
 import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
 
-@HiltAndroidApp
 class Application : Application()
