+++ app/src/main/kotlin/net/marcoromano/mooviez/app/MainActivity.kt
@@ -5,11 +5,9 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import androidx.core.view.WindowCompat
-import dagger.hilt.android.AndroidEntryPoint
 import net.marcoromano.mooviez.app.ui.AppUi
 import net.marcoromano.mooviez.app.ui.theme.AppTheme
 
-@AndroidEntryPoint
 class MainActivity : ComponentActivity() {
   override fun onCreate(savedInstanceState: Bundle?) {
     installSplashScreen()
