+++ database/fake/src/main/kotlin/net/marcoromano/mooviez/database/DatabaseFakeModule.kt
@@ -1,0 +1,9 @@
+package net.marcoromano.mooviez.database
+
+// @Module
+// @InstallIn(SingletonComponent::class)
+public object DatabaseFakeModule {
+  //  @Provides
+//  @Singleton
+  public fun database(): Database = DatabaseFake()
+}
