+++ gradle/libs.versions.toml
@@ -5,7 +5,6 @@ android-minSdk = "26"
 android-targetSdk = "33"
 androidx-activity = "1.7.2"
 androidx-compose-compiler = "1.4.7"
-androidx-hilt = "1.0.0"
 androidx-lifecycle = "2.6.1"
 androidx-navigation = "2.6.0"
 androidx-paging = "3.1.1"
@@ -14,7 +13,6 @@ androidx-testEspresso = "3.5.1"
 androidx-workManager = "2.8.1"
 coil = "2.4.0"
 google-accompanist = "0.30.1"
-google-dagger = "2.46.1"
 google-glance-tooling = "0.2.2"
 kotlin = "1.8.21"
 kotlin-inject = "0.6.1"
@@ -36,7 +34,6 @@ square-sqlDelight = "2.0.0-rc01"
 [plugins]
 android-application = { id = "com.android.application", version.ref = "plugin-android" }
 android-library = { id = "com.android.library", version.ref = "plugin-android" }
-dagger-hilt = { id = "com.google.dagger.hilt.android", version.ref = "google-dagger" }
 dependencyUpdates = { id = "com.github.ben-manes.versions", version.ref = "plugin-dependencyUpdates" }
 google-firebase-appDistribution = { id = "com.google.firebase.appdistribution", version.ref = "plugin-google-firebase-appDistribution" }
 google-firebase-crashlytics = { id = "com.google.firebase.crashlytics", version.ref = "plugin-google-firebase-crashlytics" }
@@ -44,7 +41,6 @@ google-ksp = { id = "com.google.devtools.ksp", version.ref = "plugin-google-ksp"
 google-services = { id = "com.google.gms.google-services", version.ref = "plugin-google-services" }
 kotlin-android = { id = "org.jetbrains.kotlin.android", version.ref = "kotlin" }
 kotlin-jvm = { id = "org.jetbrains.kotlin.jvm", version.ref = "kotlin" }
-kotlin-kapt = { id = "org.jetbrains.kotlin.kapt", version.ref = "kotlin" }
 kotlin-parcelize = { id = "org.jetbrains.kotlin.plugin.parcelize", version.ref = "kotlin" }
 kotlin-serialization = { id = "org.jetbrains.kotlin.plugin.serialization", version.ref = "kotlin" }
 kotlinter = { id = "org.jmailen.kotlinter", version.ref = "plugin-kotlinter" }
@@ -81,10 +77,6 @@ androidx-coreSplashscreen = "androidx.core:core-splashscreen:1.0.1"
 androidx-coreTesting = "androidx.arch.core:core-testing:2.2.0"
 androidx-datastore = "androidx.datastore:datastore:1.0.0"
 androidx-glance-glanceAppwidget = "androidx.glance:glance-appwidget:1.0.0-beta01"
-androidx-hiltCompiler = { module = "androidx.hilt:hilt-compiler", version.ref = "androidx-hilt" }
-androidx-hiltNav = { module = "androidx.hilt:hilt-navigation", version.ref = "androidx-hilt" }
-androidx-hiltNavCompose = "androidx.hilt:hilt-navigation-compose:1.0.0"
-androidx-hiltWork = { module = "androidx.hilt:hilt-work", version.ref = "androidx-hilt" }
 androidx-lifecycleCommonJava8 = { module = "androidx.lifecycle:lifecycle-common-java8", version.ref = "androidx-lifecycle" }
 androidx-lifecycleLivedata = { module = "androidx.lifecycle:lifecycle-livedata-ktx", version.ref = "androidx-lifecycle" }
 androidx-lifecycleProcess = { module = "androidx.lifecycle:lifecycle-process", version.ref = "androidx-lifecycle" }
@@ -124,12 +116,6 @@ google-accompanistPagerInd = { module = "com.google.accompanist:accompanist-page
 google-accompanistPermissions = { module = "com.google.accompanist:accompanist-permissions", version.ref = "google-accompanist" }
 google-accompanistPlaceholderMaterial = { module = "com.google.accompanist:accompanist-placeholder-material", version.ref = "google-accompanist" }
 google-accompanistSystemUi = { module = "com.google.accompanist:accompanist-systemuicontroller", version.ref = "google-accompanist" }
-google-dagger = { module = "com.google.dagger:dagger", version.ref = "google-dagger" }
-google-daggerCompiler = { module = "com.google.dagger:dagger-compiler", version.ref = "google-dagger" }
-google-daggerHiltAndroid = { module = "com.google.dagger:hilt-android", version.ref = "google-dagger" }
-google-daggerHiltCompiler = { module = "com.google.dagger:hilt-compiler", version.ref = "google-dagger" }
-google-daggerHiltCore = { module = "com.google.dagger:hilt-core", version.ref = "google-dagger" }
-google-daggerHiltTesting = { module = "com.google.dagger:hilt-android-testing", version.ref = "google-dagger" }
 google-firebase-analytics = { module = "com.google.firebase:firebase-analytics-ktx" }
 google-firebase-bom = "com.google.firebase:firebase-bom:32.1.0"
 google-firebase-config = { module = "com.google.firebase:firebase-config-ktx" }
@@ -155,7 +141,6 @@ kotlinx-serialization-core = { module = "org.jetbrains.kotlinx:kotlinx-serializa
 kotlinx-serialization-json = { module = "org.jetbrains.kotlinx:kotlinx-serialization-json", version.ref = "kotlinx-serialization" }
 kotlinx-serialization-protobuf = { module = "org.jetbrains.kotlinx:kotlinx-serialization-protobuf", version.ref = "kotlinx-serialization" }
 plugin-android = { module = "com.android.tools.build:gradle", version.ref = "plugin-android" }
-plugin-dagger-hilt = { module = "com.google.dagger:hilt-android-gradle-plugin", version.ref = "google-dagger" }
 plugin-dependencyUpdates = { module = "com.github.ben-manes:gradle-versions-plugin", version.ref = "plugin-dependencyUpdates" }
 plugin-google-firebase-appDistribution = { module = "com.google.firebase:firebase-appdistribution-gradle", version.ref = "plugin-google-firebase-appDistribution" }
 plugin-google-firebase-crashlytics = { module = "com.google.firebase:firebase-crashlytics-gradle", version.ref = "plugin-google-firebase-crashlytics" }
