+++ httpapi/fake/src/main/kotlin/net/marcoromano/mooviez/httpapi/HttpApiFakeModule.kt
@@ -1,0 +1,9 @@
+package net.marcoromano.mooviez.httpapi
+
+// @Module
+// @InstallIn(SingletonComponent::class)
+public class HttpApiFakeModule {
+  //  @Provides
+//  @Singleton
+  public fun httpApi(): HttpApi = HttpApiFake()
+}
